refactor: separate event handling logics from the controller as a service
