+++ controllers/rtc-exchanger/index.js
@@ -1,5 +1,3 @@
 module.exports = {
   ...require('./rtc-exchanger.controller'),
-  ...require('./socket-stream-mapper'),
-  ...require('./socket-user-mapper'),
 }
