+++ services/rtc-exchanger/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  ...require('./rtc-exchanger.service'),
+  ...require('./socket-stream-mapper'),
+  ...require('./socket-user-mapper'),
+}
